Create ThingReference class (WIP)
In Python 3, we can't override the __class__ object inside
Thing._getdata (not entirely sure why). As an alternative, try to
separate the cases where a Thing is just a shallow reference without
data and when it's been populated with information from the database.

This causes loads of things to breaks, and needs way more testing.